BLE: Move generic implementation and headers into source/generic
