+++ connectivity/FEATURE_BLE/source/generic/GattClientImpl.h
