+++ connectivity/FEATURE_BLE/source/generic/GattServerImpl.cpp
@@ -18,7 +18,7 @@
 
 #include "BLERoles.h"
 #include <algorithm>
-#include "ble/internal/GattServerImpl.h"
+#include "GattServerImpl.h"
 #include "ble/internal/BLEInstanceBase.h"
 #include "wsf_types.h"
 #include "att_api.h"
