+++ connectivity/FEATURE_BLE/source/generic/SecurityManagerImpl.h
